init style refac. TODO - sidenav flightbadge margins
