+++ src/components/flightBadge/FlightBadge.scss
@@ -5,7 +5,7 @@
   background: #009fe3;
   border: 1px solid white;
   border-radius: 0.25rem;
-  margin: 0px -10px;
+  margin: 0px 5px 0px -10px;
   padding: 10px 5px;
 
   .flight-number {
