+++ src/components/sidenavContainer/SidenavContainer.scss
@@ -2,7 +2,8 @@
 @import "~bootstrap/scss/variables";
 @import "~bootstrap/scss/mixins";
 
-$background-image-url: "../../images/plane01.jpg";
+// $background-image-url: "../../images/sofy-m-halftoneco-N7uekwIyQ7E-unsplash.jpg";
+$background-image-url: "../../images/chairs.jpg";
 $wco-blue: #0082c3;
 $wco-blue-dark: darken($wco-blue, 20%);
 $wco-gray: rgb(141, 144, 147);
@@ -114,7 +115,7 @@ $transparent: #0082c3d9;
 
 .filter-fg,
 .filter-fg-open {
-  padding-right: 15px;
+  // padding-right: 15px;
   color: #f8f9fa;
   z-index: 10;
 }
@@ -131,9 +132,11 @@ $transparent: #0082c3d9;
 
 .filter-bg {
   z-index: 5;
-  background: linear-gradient(#40403f, #40403fe8, #40403f99),
+  // background: linear-gradient(#40403f, #40403fe8, #40403f99),
+  //   url($background-image-url) no-repeat no-repeat fixed;
+  background: linear-gradient(#40403f, #40403fdd, #40403f55),
     url($background-image-url) no-repeat no-repeat fixed;
-  background-position: -600px 0px;
+  // background-position: -600px 0px;
   background-size: cover;
 }
 
@@ -167,5 +170,8 @@ hr {
   line-break: anywhere;
 }
 
-.item-renderer > .input {
+.filterform-container > form {
+  background-color: #40404099;
+  padding: 20px;
+  border-radius: 0.25rem;
 }
