+++ src/pages/login/ForgotPassword.js
@@ -2,8 +2,9 @@ import React, { useState } from "react";
 import Form from "../../components/form/Form";
 import LabelledInput from "../../components/labelledInput/LabelledInput";
 import { forgotPassword } from "../../services/serviceWrapper";
-import { Container, Alert } from "react-bootstrap";
+import { Container, Alert, Col } from "react-bootstrap";
 import Title from "../../components/title/Title";
+import Main from "../../components/main/Main";
 import Xl8 from "../../components/xl8/Xl8";
 import { hasData } from "../../utils/utils";
 import { Link } from "@reach/router";
@@ -27,43 +28,49 @@ const ForgotPassword = props => {
   };
 
   return (
-    <Container className="password-reset-container" fluid>
-      <Title title={<Xl8 xid="fopa001">Reset Password</Xl8>} uri={props.uri} />
-
-      {resetLinkSent ? (
-        <>
-          <Alert variant="success">
-            A password reset link is sent to your mail. Please follow the instructions on
-            the email to reset your password.
-          </Alert>
-          <Link to="/login">Login to GTAS</Link>
-        </>
-      ) : (
-        <>
-          {displayErrorMessage && <Alert variant="danger">{errorMessage}</Alert>}
-          <Form
-            submitService={forgotPassword.post}
-            title=""
-            callback={forgotPasswordCallback}
-            action="add"
-            redirectTo="/login"
-            cancellable
-          >
-            <LabelledInput
-              datafield
-              labelText="User Id"
-              inputType="text"
-              name="userId"
-              required={true}
-              inputVal=""
-              alt="nothing"
-              callback={cb}
-              spacebetween
-            />
-          </Form>
-        </>
-      )}
-    </Container>
+    <Main className="unauthed bg-image">
+      <Container className="login d-flex align-items-center py-5" fluid>
+        <Col className="unauthed-form offset-3 col-4">
+          {resetLinkSent ? (
+            <>
+              <Alert variant="success">
+                A password reset link is sent to your mail. Please follow the instructions
+                on the email to reset your password.
+              </Alert>
+              <Link to="/login">Login to GTAS</Link>
+            </>
+          ) : (
+            <>
+              {displayErrorMessage && <Alert variant="danger">{errorMessage}</Alert>}
+              <div className="text-center h3">
+                <Xl8 xid="fopa001">Reset Password</Xl8>
+              </div>
+              <br />
+              <Form
+                submitService={forgotPassword.post}
+                title=""
+                callback={forgotPasswordCallback}
+                action="add"
+                redirectTo="/login"
+                cancellable
+              >
+                <LabelledInput
+                  datafield
+                  labelText="User Id"
+                  inputType="text"
+                  name="userId"
+                  required={true}
+                  inputVal=""
+                  alt="nothing"
+                  callback={cb}
+                  spacebetween
+                />
+              </Form>
+            </>
+          )}
+        </Col>
+      </Container>
+    </Main>
   );
 };
