--- src/pages/signUp/SignUp.scss
@@ -1,15 +1,0 @@
-$transparent: rgba(223, 231, 247, 0.14);
-
-.sign-up-container {
-  background: linear-gradient($transparent, #e5edf485),
-    url("../../images/WCO_background1000.png") no-repeat center center fixed;
-  background-size: cover;
-  min-height: 100vh;
-  margin-bottom: 0em;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  flex-direction: column;
-  background-blend-mode: overlay;
-}
